TIP-843: Move writer to enrichment
