+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Writer/Database/ProductModelDescendantsWriter.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Pim\Component\Connector\Writer\Database;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Writer\Database;
 
 use Akeneo\Tool\Component\Batch\Item\ItemWriterInterface;
 use Akeneo\Tool\Component\Batch\Model\StepExecution;
