+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Writer/File/Xlsx/ProductWriter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\Connector\Writer\File\Csv;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Writer\File\Xlsx;
 
 use Akeneo\Tool\Component\Batch\Item\FlushableInterface;
 use Akeneo\Tool\Component\Batch\Item\InitializableInterface;
@@ -10,10 +10,10 @@ use Pim\Component\Connector\Writer\File\AbstractItemMediaWriter;
 use Pim\Component\Connector\Writer\File\ArchivableWriterInterface;
 
 /**
- * Write product data into a csv file on the local filesystem
+ * Write product data into a XLSX file on the local filesystem
  *
- * @author    Yohan Blain <yohan.blain@akeneo.com>
- * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
+ * @author    Arnaud Langlade <arnaud.langlade@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 class ProductWriter extends AbstractItemMediaWriter implements
@@ -28,14 +28,7 @@ class ProductWriter extends AbstractItemMediaWriter implements
      */
     protected function getWriterConfiguration()
     {
-        $parameters = $this->stepExecution->getJobParameters();
-
-        return [
-            'type'           => 'csv',
-            'fieldDelimiter' => $parameters->get('delimiter'),
-            'fieldEnclosure' => $parameters->get('enclosure'),
-            'shouldAddBOM'   => false,
-        ];
+        return ['type' => 'xlsx'];
     }
 
     /**
