+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Writer/Database/ProductModelDescendantsWriterSpec.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Writer\Database;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Writer\Database;
 
 use Akeneo\Tool\Component\Batch\Model\StepExecution;
 use Akeneo\Tool\Component\StorageUtils\Cache\EntityManagerClearerInterface;
 use Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
-use Pim\Component\Connector\Writer\Database\ProductModelDescendantsWriter;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\Writer\Database\ProductModelDescendantsWriter;
 use PhpSpec\ObjectBehavior;
 
 class ProductModelDescendantsWriterSpec extends ObjectBehavior
